Organize the methods of TSO dispatcher
